+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - improved support when using subfolder of git root as project directory
 - properly loading messages from the last session on startup
 - added None reasoning effort option for OpenAI provider
+- added dynamic model loading for LLM providers
 
 ## [0.30.0]